dashboard: update cloudfront settigns to actually display things correctly
Matching the connection to the backend as the rest of the load balanced
ECS setup does things, though might be useful to have further improvements
added later (e.g. currently TLS is terminated on Cloudfront, and after that
it's HTTP traffic inside AWS. Ideally all traffic should be HTTPS,
but that will need some load balancer and container changes that are
still to be determined.

Signed-off-by: Gergely Imreh <gergely.imreh@faculty.ai>