dashboard: remove caching for the time being
While it is no doubt handy some times, in our particular setup,
with dynamic pages, it is not the most useful or effective way
of doing things as it wouldn't cache much anyways. So to remove
complexity, remove caching for now as well. In the future we
would know how to add it back anyways.